SqlNameHelper tests
