+++ src/ServiceControl.Monitoring.SmokeTests.SQLServer/ServiceControl.Monitoring.SmokeTests.SQLServer.csproj
@@ -78,6 +78,7 @@
     <Compile Include="NServiceBusAcceptanceTest.cs" />
     <Compile Include="ScenarioDescriptors\EnvironmentHelper.cs" />
     <Compile Include="Tests\ApiIntegrationTest.cs" />
+    <Compile Include="Tests\SqNameHelperTests.cs" />
     <Compile Include="Tests\When_querying_queue_length.cs" />
     <Compile Include="Tests\When_querying_retries_data.cs" />
     <Compile Include="Tests\When_querying_timings_data.cs" />
