+++ src/ServiceControl.Transports.SQLServer/InternalsVisibleTo.cs
@@ -1,0 +1,3 @@
+﻿using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("ServiceControl.Monitoring.SmokeTests.SQLServer")]
